find where wrong ssim comes from, trying to fix it.
